✏️ Rename: component `ContentsTable` to `ContentTable`
